+++ client/index.js
@@ -1,39 +1,39 @@
+// @ts-check
+/** Type Definitions */
+/** @module rill/http */
+/** @typedef {(req: IncomingMessage, res: ServerResponse?) => any} RequestHandler */
 'use strict'
 
 var Server = require('./server')
 var IncomingMessage = require('./incoming-message')
 var ServerResponse = require('./server-response')
+// @ts-ignore
 var STATUS_CODES = require('statuses/codes.json')
 
-/** @type {object} */
-exports.STATUS_CODES = STATUS_CODES
-
-/** @type {string[]} */
-exports.METHODS = [
-  'OPTIONS',
-  'HEAD',
-  'GET',
-  'PUT',
-  'POST',
-  'PATCH',
-  'DELETE'
-]
-
-/** @type {Server} */
-exports.Server = Server
-
-/** @type {IncomingMessage} */
-exports.IncomingMessage = IncomingMessage
-
-/** @type {ServerResponse} */
-exports.ServerResponse = ServerResponse
-
-/**
- * Creates a new mock http server in the browser.
- *
- * @return {Server}
- */
-exports.createServer = function () {
-  var onRequest = arguments[arguments.length - 1]
-  return new Server(onRequest)
+module.exports = {
+  Server: Server,
+  IncomingMessage: IncomingMessage,
+  ServerResponse: ServerResponse,
+  /** @type {object} */
+  STATUS_CODES: STATUS_CODES,
+  /** @type {string[]} */
+  METHODS: [
+    'OPTIONS',
+    'HEAD',
+    'GET',
+    'PUT',
+    'POST',
+    'PATCH',
+    'DELETE'
+  ],
+  /**
+   * Creates a new mock http server in the browser.
+   * Any arguments provided (besides the last one) are ignored in the browser.
+   *
+   * @param {...RequestHandler|{[x: string]: any}} [onRequest] - A function called on each request.
+   * @return {Server}
+   */
+  createServer: function createServer (onRequest) {
+    return new Server(arguments[arguments.length - 1])
+  }
 }