+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 6.0.0 - 2017-10-10
+### Changed
+- Removes and moves some hidden `_properties` from `IncomingMessage`. (Requires an update to @rill/static).
+- Better support for typescript with improved JSDoc.
+- Moves initialization of non-methods to the constructor instead of the prototype for server-response.
+- Update dev dependencies.
+
 ## 5.0.2 - 2017-05-18
 ### Changed
 - Update dev dependencies.